Refine filter handling into dedicated applier
